requirements.txt
django==3.2
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
requests~=2.28.0
geopy~=2.2.0
numpy~=1.23.0
django-phonenumber-field[phonenumbers]~=6.3.0
djangorestframework~=3.13.0
django-extensions~=3.2.0
